pyproject.toml
[project]
name = "mcp-servers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.57.1",
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "mcp[cli]>=1.10.1",
    "openai>=1.93.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.1",
]
